Decoupled Gameloop
loops are interfaces and separated to make it easier to add new ones at different tick rates
