+++ ArchaicQuestII.GameLogic/Core/IGameLoop.cs
@@ -8,16 +8,6 @@ namespace ArchaicQuestII.GameLogic.Core
 {
     public interface IGameLoop
     {
-        // HP, Mana, Moves gain
-        int GainAmount(int value, Player player);
-        Task UpdateTime();
-        Task UpdateRoomEmote();
-        Task UpdateMobEmote();
-        Task UpdatePlayers();
-        Task UpdatePlayerLag();
-        Task UpdateCombat();
-        Task UpdateWorldTime();
-        Task Tick();
-        void IdleCheck(Player player);
+        void StartLoops();
     }
 }
